% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_chunkwise}
\alias{read_chunkwise}
\title{Read text file \code{filepath} in chunks of \code{chunk_size} lines and pass each
chunk to a user-provided \code{handler} function.}
\usage{
read_chunkwise(
  filepath,
  handler,
  chunk_size = 100000L,
  repeat_header = TRUE,
  header_rows = 1L,
  skip = 0L,
  max_rows = NULL
)
}
\arguments{
\item{filepath}{Path to the file to read}

\item{handler}{User-provided handler function to which chunks will be passed}

\item{chunk_size}{number of lines to read in as one chunk (defaults to 100000)}

\item{repeat_header}{Append header line(s) to beginning of each chunk
(defaults to TRUE)}

\item{header_rows}{number of lines at the beginning of file to consider as header}

\item{skip}{number of lines at the beginning of the file to skip (defaults to 0)}

\item{max_rows}{If not \code{NULL}, only read the first \code{max_rows} lines of data}
}
\description{
Read text file \code{filepath} in chunks of \code{chunk_size} lines and pass each
chunk to a user-provided \code{handler} function.
}
