% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chunkwise_dt.R
\name{read_chunkwise_dt}
\alias{read_chunkwise_dt}
\title{Read in a large tabular text file as a data.table, chunk by chunk}
\usage{
read_chunkwise_dt(file, filter = identity, ...)
}
\arguments{
\item{file}{Path to the text file to read}

\item{filter}{Filter function (defaults to \code{identity})}

\item{...}{Parameters passed through to \code{read_chunkwise}}
}
\value{
A \code{data.table} generated from the data in the text file
}
\description{
Crucially, a custom \code{filter} function can be provided to perform arbitrary
transformations on each chunk before appending it to the full data.table.
}
